Floor the rating for fedmsg messages.
This, like in fedora-infra/fedmsg#201, is just to reduce erroneous
fedmsg validation failures.   Without this, the rating gets encoded on
one machines as 0.9111111111112 on one machine and as 0.91111111111 on
another.  Those are not the same strings, and so the signature
computation produces different signatures -- recipients think that the
message is "invalid".